requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==2.5.1
Flask-Login==0.6.2
Werkzeug==2.3.7
SQLAlchemy==1.4.41
gunicorn==21.2.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0
